Refactor router for user and test
