+++ backend/tests/router/test_router_user.py
@@ -8,7 +8,7 @@ endPoint = f"{apiPrefix}/fam_users"
 
 
 def test_get_fam_users_nodata(testClient_fixture):
-    response = testClient_fixture.get(endPoint)
+    response = testClient_fixture.get(f"{endPoint}/")
     LOGGER.debug(f"endPoint: {endPoint}")
     LOGGER.debug(f"response {response}")
     assert response.status_code == 200
@@ -18,7 +18,7 @@ def test_get_fam_users_nodata(testClient_fixture):
 
 
 def test_get_fam_users_withdata(user_client_withUsers):
-    response = user_client_withUsers.get(endPoint)
+    response = user_client_withUsers.get(f"{endPoint}/")
     LOGGER.debug(f"response: {response}")
     data = response.json()
     LOGGER.debug(f"data: {data}")
@@ -54,7 +54,7 @@ def test_get_fam_user(user_client_withUsers):
     # retrieval of user is by id, don't know what id got
     # assigned so getting all the users, from that can
     # test the get individual user end point
-    allUsersResponse = user_client_withUsers.get(endPoint)
+    allUsersResponse = user_client_withUsers.get(f"{endPoint}/")
     allUsersData = allUsersResponse.json()
     LOGGER.debug(f"user data: {allUsersData}")
     for user in allUsersData:
@@ -70,7 +70,7 @@ def test_post_fam_users(testClient_fixture, testUserData):
     testUserData['create_date'] = str(testUserData['create_date'])
     testUserData['update_date'] = str(testUserData['update_date'])
 
-    resp = testClient_fixture.post(endPoint, json=testUserData)
+    resp = testClient_fixture.post(f"{endPoint}/", json=testUserData)
     body = resp.json()
     LOGGER.debug(f"body: {body}")
     assert resp.status_code == 422
@@ -80,7 +80,7 @@ def test_post_fam_users(testClient_fixture, testUserData):
 
     # fix the data so the post should succeed
     testUserData["user_type"] = "z"
-    resp = testClient_fixture.post(endPoint, json=testUserData)
+    resp = testClient_fixture.post(f"{endPoint}/", json=testUserData)
     respData = resp.json()
     LOGGER.debug(f"resp data: {resp.json()}")
     assert resp.status_code == 200
@@ -88,7 +88,7 @@ def test_post_fam_users(testClient_fixture, testUserData):
     # verify that a record with the same guid cannot be entered
     respDataCopy = respData.copy()
     del respDataCopy["user_id"]
-    resp = testClient_fixture.post(endPoint, json=respDataCopy)
+    resp = testClient_fixture.post(f"{endPoint}/", json=respDataCopy)
     respBody = resp.json()
     LOGGER.debug(f"resp data: {resp.status_code} - {resp.json()}")
     assert resp.status_code == 422
@@ -114,7 +114,7 @@ def user_client_withUsersNoCleanup(testClient_fixture, testUserData):
     testUserData['create_date'] = str(testUserData['create_date'])
     testUserData['update_date'] = str(testUserData['update_date'])
 
-    resp = testClient_fixture.post(endPoint, json=testUserData)
+    resp = testClient_fixture.post(f"{endPoint}/", json=testUserData)
     LOGGER.debug(f"setup user table with data: {resp.status_code} " +
                  f"{resp.reason}")
     LOGGER.debug(f"setup data: {testUserData}")
@@ -131,7 +131,7 @@ def user_client_withUsers(testClient_fixture, testUserData):
     testUserData['create_date'] = str(testUserData['create_date'])
     testUserData['update_date'] = str(testUserData['update_date'])
 
-    resp = testClient_fixture.post(endPoint, json=testUserData)
+    resp = testClient_fixture.post(f"{endPoint}/", json=testUserData)
     LOGGER.debug(f"setup user table with data: {resp.status_code}" +
                  f"  {resp.reason}")
     if resp.status_code != 200:
